+++ src/PHPHtmlParser/Dom/Tag.php
@@ -53,6 +53,11 @@ class Tag
     protected $encode = null;
 
     /**
+     * @var bool
+     */
+    private $HtmlSpecialCharsDecode = false;
+
+    /**
      * Sets up the tag with a name.
      *
      * @param $name
@@ -143,6 +148,15 @@ class Tag
     }
 
     /**
+     * @param bool $htmlSpecialCharsDecode
+     * @return void
+     */
+    public function setHtmlSpecialCharsDecode($htmlSpecialCharsDecode = false): void
+    {
+        $this->HtmlSpecialCharsDecode = $htmlSpecialCharsDecode;
+    }
+
+    /**
      * Sets the noise for this tag (if any)
      *
      * @param string $noise
@@ -173,6 +187,9 @@ class Tag
                 'doubleQuote' => true,
             ];
         }
+        if ($this->HtmlSpecialCharsDecode) {
+            $value['value'] = htmlspecialchars_decode($value['value']);
+        }
         $this->attr[$key] = $value;
 
         return $this;
@@ -186,7 +203,6 @@ class Tag
      */
     public function setStyleAttributeValue($attr_key, $attr_value): void
     {
-
         $style_array = $this->getStyleAttributeArray();
         $style_array[$attr_key] = $attr_value;