+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new option `depthFirstSearch`.
 - Deprecated option `depthFirstSearch` and marked for removal in `3.0.0`.
 - Added multi class selections support.
-- Added case insensitive attribute matching
+- Added case insensitive attribute matching.
+- Added new option `htmlSpecialCharsDecode`.
 
 ### Changed
 - Started using a changelog.